A refactoring of slug creation logic
Previous logic created a problem with numeric titles
so we refactored the logic to this new one
this has 2 problems, one is that it may creat a n+1 problem
second is that if i create a thread, then delete it,and some one create
another with same title, if someone else link to my thread, it will be
redirected to second one.

so implementing a timestamp or id approuch would be better.